add partnersetupcard to setuppage
