+++ src/components/onboarding/OnboardingChecklistCard.module.scss
