+++ src/components/onboarding/OnboardingChecklistCard.tsx
